Helfermethode intsToNotes(ArrayList<Integer>[][], ArrayList<Integer>[][] aus MelodyGenerator in neue Helferklasse MelodyHelper ausgelagert
